update caculate price

+++ src/main/java/com/thoughtworks/iamcoach/vo/PromotionSecondHalf.java
@@ -3,6 +3,6 @@ package com.thoughtworks.iamcoach.vo;
 public class PromotionSecondHalf extends Promotion {
     @Override
     public double caculatePromotionPrice(Item item, double num) {
-        return num / 2 * (item.getPrice() / 2.0);
+        return (int)num / 2 * (item.getPrice() / 2.0);
     }
 }
