+++ src/main/java/com/thoughtworks/iamcoach/vo/PromotionTwoForOne.java
@@ -3,6 +3,6 @@ package com.thoughtworks.iamcoach.vo;
 public class PromotionTwoForOne extends Promotion {
     @Override
     public double caculatePromotionPrice(Item item, double num) {
-        return num / 3 * item.getPrice();
+        return (int)num / 3 * item.getPrice();
     }
 }
